Ngrok and Expo
Got ngrok to tunnel frontend and api for expo app. Added api.ts to app folder in mobile directory to fix an issue with the vite port causing a 404 error in expo. This is to avoid altering Program.cs. It still deafaults to the vite port about half the time. Still needs debugging.