Connected mobile to backend api.
Configured Krestel to open it up to my local network. Added CORS policy to allow any origin. This is to bypass a lot of certificate issues when testing through a mobile emmulator. app.UseHttpsRedirection() has to be commented out in Program.cs during mobile testing or it will redirect to https and block api requests from mobile. Servies for api requests are in app/services: Client.ts (handels actual HTTP comminication with the backend), index.ts (has a default export that combines all API services for easy importing elsewhere in the app). Took a modular approach with so each domain area has it's own service file (ex moviesApi.ts). This is to make the app easy to scale since we have a lot of features planned and a lot of endpoints. Added http://0.0.0.0:5249 to launchSettings.json to bind the API server to all network interfaces instead of just localhost. This allows external devices like Android emulator to connect to the API. Without this, the API can only be reached by the local host.