Navbars and navigation to theaters and movies.
Got the app to open to movies screen. Moved navigation folder to the top level of the mobile folder for expo router. Removed the theaters button from the movies screen and applied that navigation to the "Locations" button in the bottom NavBar. So far the the movies and theaters navigation works.