Mobile Movies layout
Added themeToggle back to Web so I can see the web version and made a small edit to line 99 of the NavBar.tsx in web directory. Turned movie mobile layout into a grid. The only console error for fetching movie posters on the movies screen. Working on getting that to fetch static files.