Mobile Movie Landing page
Finished up the movies page. Refractored how it fetches movie posters and what details it shows on the movie card. Tested in browser and on android mobile, and it works. Changed the layout to a grid. Saved movie posters in Selu383.SP25.P03.Mobile\assets\images\posters becasue fetching them from the backend created dependency issues and this is a quicker fix.